moves the map layers to separate files
