+++ src/components/Map/index.js
@@ -1,11 +1,7 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 
 import 'services/map'; // loads L.Proj (Proj binding leaflet)
-import { CATEGORY_NAMES } from 'shared/configuration/categories';
-import { fetchCameraAreas } from 'services/layer-aggregator/layersFetcher';
-import LAYERS_CONFIG from 'services/layer-aggregator/layersConfig';
-import layersReader from 'services/layer-aggregator/layersReader';
 import MapLegend from '../MapLegend';
 import DeviceDetails from '../DeviceDetails';
 import CameraAreaDetails from '../CameraAreaDetails';
@@ -15,69 +11,8 @@ import './amaps-style.scss';
 import useMap from './hooks/useMap';
 import { MapContainerStyle } from './MapStyle';
 import useLayerManager from './hooks/useLayerManager';
-
-const CamerasLayer = ({ map, data: cameras, selectLayerItem, addLayerData, removeLayerData, layerManager }) => {
-  const { addPolygonLayer, removePolygonLayer } = layerManager;
-
-  const showCameraAreaDetail = item => {
-    selectLayerItem('cameras', item);
-  };
-
-  useEffect(() => {
-    if (map!== null && cameras != null) {
-      addPolygonLayer(CATEGORY_NAMES.CAMERA_TOEZICHTSGEBIED, cameras, showCameraAreaDetail);
-    }
-    return () => removePolygonLayer();
-  }, [map, cameras]);
-
-  useEffect(() => {
-    (async () => {
-      const features = await fetchCameraAreas();
-
-      addLayerData('cameras', features);
-    })();
-
-    return () => {
-      removeLayerData('cameras')
-    };
-
-  }, []);
-
-  return null;
-};
-
-const ClusterLayer = ({ map, data: devices, selectLayerItem, addLayerData, removeLayerData,layerManager }) => {
-  const { addPointClusterLayer, removeClusterPointLayer } = layerManager;
-
-  const showDeviceDetail = device => {
-    if (device) {
-      selectLayerItem('devices', device);
-    } else {
-      selectLayerItem();
-    }
-  };
-
-  useEffect(() => {
-    if (map!== null && devices != null) {
-      addPointClusterLayer(devices.features, showDeviceDetail);
-    }
-    return () => removeClusterPointLayer();
-  }, [map, devices]);
-
-  useEffect(() => {
-    (async () => {
-      const results = await layersReader(LAYERS_CONFIG);
-      const features = results.reduce((acc, { layer }) => [...acc, ...layer.features], []);
-
-      addLayerData('devices', { type: 'FeatureCollection', name: 'devices', features });
-    })();
-    return () => {
-      removeLayerData('devices')
-    };
-  }, []);
-
-  return null;
-};
+import DevicesLayer from './components/DevicesLayer';
+import CamerasLayer from './components/CamerasLayer';
 
 const Map = ({ layers, selectedLayer, selectedItem, addLayerData, removeLayerData, selectLayerItem }) => {
   const mapRef = useMap();
@@ -87,14 +22,27 @@ const Map = ({ layers, selectedLayer, selectedItem, addLayerData, removeLayerDat
     selectLayerItem();
   };
 
-
   return (
     <MapContainerStyle className="map-component">
       <div className="map">
         <div id="mapdiv">
           <MapLegend onToggleCategory={name => layerManager.toggleLayer(name)} />
-          <ClusterLayer map={mapRef.current} data={layers.devices} addLayerData={addLayerData} removeLayerData={removeLayerData} selectLayerItem={selectLayerItem} layerManager={layerManager}/>
-          <CamerasLayer map={mapRef.current} data={layers.cameras} addLayerData={addLayerData} removeLayerData={removeLayerData} selectLayerItem={selectLayerItem} layerManager={layerManager}/>
+          <DevicesLayer
+            map={mapRef.current}
+            data={layers.devices}
+            addLayerData={addLayerData}
+            removeLayerData={removeLayerData}
+            selectLayerItem={selectLayerItem}
+            layerManager={layerManager}
+          />
+          <CamerasLayer
+            map={mapRef.current}
+            data={layers.cameras}
+            addLayerData={addLayerData}
+            removeLayerData={removeLayerData}
+            selectLayerItem={selectLayerItem}
+            layerManager={layerManager}
+          />
           {selectedLayer === 'devices' && <DeviceDetails device={selectedItem} onDeviceDetailsClose={clearSelection} />}
 
           {selectedLayer === 'cameras' && <CameraAreaDetails onDeviceDetailsClose={clearSelection} />}
